Separate bookworm state
